split up request-validation-schema into multiple modules
