+++ server/app/routes/suggestion-router.ts
@@ -2,7 +2,7 @@ import { Router } from "express";
 
 const router = Router();
 import suggestionController from "../controllers/suggestion-controller";
-import { suggestionPutRequestSchema, suggestionPostRequestSchema } from '../request-validation-schema'
+import { suggestionPutRequestSchema, suggestionPostRequestSchema } from '../validation-schema/suggestion-schema'
 import { requestValidationMiddleware } from '../../middleware/request-validation-middlewares'
 
 router.get("/", suggestionController.getAll);
