+++ server/app/routes/tag-router.ts
@@ -2,8 +2,8 @@ import { Router } from "express";
 import * as tagController from "../controllers/tag-controller";
 import jwtSession from "../../middleware/jwt-session";
 import {requestValidationMiddleware} from '../../middleware/request-validation-middlewares'
-import { tagPutRequestSchema } from "../request-validation-schema";
-import { tagPostRequestSchema } from "../request-validation-schema";
+import { tagPutRequestSchema } from "../validation-schema/tag-schema";
+import { tagPostRequestSchema } from "../validation-schema/tag-schema";
 
 const router = Router();
