+++ server/app/routes/tenant-router.ts
@@ -2,7 +2,7 @@ import { Router } from "express";
 const router = Router();
 import tenantController from "../controllers/tenant-controller";
 import { requestValidationMiddleware } from '../../middleware/request-validation-middlewares';
-import { tenantRequestSchema } from '../request-validation-schema';
+import { tenantRequestSchema } from '../validation-schema/tenant-schema';
 
 router.get("/", tenantController.getAll);
 router.get("/:id", tenantController.getById);
