+++ server/app/validation-schema/suggestion-schema.ts
@@ -1,71 +1,5 @@
 import { JSONSchemaType } from "ajv";
-import { Tenant } from '../types/tenant-types';
-import { StakeholderTag } from '../types/tag-types';
-import { Suggestion } from "../types/suggestion-types";
-
-export const tenantRequestSchema: JSONSchemaType<Tenant> = {
-  type: "object",
-  required: ["id", "name", "code"],
-  properties: {
-    id: {
-      type: "integer",
-      minimum: 1,
-      maximum: 10000,
-    },
-    name: {
-      type: "string",
-      minLength: 1,
-      maxLength: 256,
-    },
-    code: {
-      type: "string",
-      minLength: 1,
-      maxLength: 256,
-    },
-  },
-  additionalProperties: false,
-};
-
-export const tagPutRequestSchema: JSONSchemaType<StakeholderTag> = {
-  type: "object",
-  required: ["id", "name", "tenantId"],
-  properties: {
-    id: {
-      type: "integer",
-      minimum: 1,
-      maximum: 10000,
-    },
-    name: {
-      type: "string",
-      minLength: 1,
-      maxLength: 256,
-    },
-    tenantId: {
-      type: "integer",
-      minimum: 1,
-      maximum: 10000
-    },
-  },
-  additionalProperties: false,
-};
-
-export const tagPostRequestSchema: JSONSchemaType<Omit<StakeholderTag, "id">> = {
-  type: "object",
-  required: ["name", "tenantId"],
-  properties: {
-    name: {
-      type: "string",
-      minLength: 1,
-      maxLength: 256
-    },
-    tenantId: {
-      type: "integer",
-      minimum: 1,
-      maximum: 10000,
-    },
-  },
-  additionalProperties: false
-}
+import { Suggestion } from "../../types/suggestion-types";
 
 export const suggestionPutRequestSchema: JSONSchemaType<Suggestion> = {
   type: "object",
